Added decoder + [filter class]
Added a filter class, which tokenizes the filter entry, checks each part of it and removes invalid filters.
Added a decoder script with primary public 'decode' function, which uses an dictionary handler for calling the fitting decoding functions for the packet, in the future this will replace the Scapy "decoding" that i have now.

# To Do:
- Create a new script with primary function that gets the information frame and the packet itself, and sends the packet and the information frame to the fitting function to create labels for all of the packet's information.
- Try to remove Scapy from the project, sniff the packets myself with something i myself coded